+++ assets/js/components/tracks/annotation_tracks.ts
@@ -1,0 +1,67 @@
+import { removeChildren } from "../../track/utils";
+import { BandTrack } from "./band_track";
+
+const template = document.createElement("template");
+template.innerHTML = String.raw`
+    <div id="container"></div>
+`;
+
+export class AnnotationTracks extends HTMLElement {
+  protected _root: ShadowRoot;
+  public label: string = "Annotation tracks";
+
+  trackHeight: number;
+  getRenderData: () => Promise<AnnotationTracksData>;
+  renderData: AnnotationTracksData | null;
+
+  parentContainer: HTMLDivElement;
+
+  constructor() {
+    super();
+  }
+
+  connectedCallback() {
+    this._root = this.attachShadow({ mode: "open" });
+    this._root.appendChild(template.content.cloneNode(true));
+
+    this.parentContainer = this._root.getElementById(
+      "container",
+    ) as HTMLDivElement;
+  }
+
+  async initialize(
+    trackHeight: number,
+    getRenderData: () => Promise<AnnotationTracksData>,
+  ) {
+    this.trackHeight = trackHeight;
+    this.getRenderData = getRenderData;
+  }
+
+  async render(updateData: boolean) {
+    if (updateData || this.renderData == null) {
+      this.renderData = await this.getRenderData();
+    }
+
+    removeChildren(this.parentContainer);
+
+    const { xRange, annotations } = this.renderData;
+
+    for (const annotData of annotations) {
+      const annotTrack = new BandTrack();
+      this.parentContainer.appendChild(annotTrack);
+      await annotTrack.initialize(
+        `Annot: ${annotData}`,
+        this.trackHeight,
+        async () => {
+          return {
+            bands: annotData,
+            xRange,
+          };
+        },
+      );
+      await annotTrack.render(true);
+    }
+  }
+}
+
+customElements.define("annotation-tracks", AnnotationTracks);