+++ assets/js/components/tracks/dot_track.ts
@@ -1,3 +1,4 @@
+import { STYLE } from "../../util/constants";
 import { CanvasTrack } from "./canvas_track";
 import {
   drawDotsScaled,
@@ -7,28 +8,34 @@ import {
 
 export class DotTrack extends CanvasTrack {
   renderData: DotTrackData | null;
+  getRenderData: () => Promise<DotTrackData>;
+  yRange: Rng;
 
-  initialize(label: string, trackHeight: number) {
-    super.initializeCanvas(label, trackHeight);
+  async initialize(label: string, trackHeight: number, yRange: Rng, getRenderData: () => Promise<DotTrackData>) {
+    super.initializeCanvas(label, trackHeight, false);
+    this.getRenderData = getRenderData;
+    this.yRange = yRange;
+    await this.updateRenderData();
   }
 
-  updateRenderData(renderData: DotTrackData) {
-    this.renderData = renderData;
+  async updateRenderData() {
+    this.renderData = await this.getRenderData();
   }
 
-  render() {
-    if (this.renderData == null) {
-      throw Error(`Render data should be assigned before render (current track: ${this.label})`);
+  async render(updateData: boolean) {
+    if (updateData || this.renderData == null) {
+      this.renderData = await this.getRenderData();
     }
 
-    const { xRange, yRange, dots } = this.renderData;
+    const { xRange, dots } = this.renderData;
+    const yRange = this.yRange;
 
     super.syncDimensions();
 
     const xScale = getLinearScale(xRange, [0, this.dimensions.width]);
     const yScale = getLinearScale(yRange, [0, this.dimensions.height]);
 
-    renderBorder(this.ctx, this.dimensions);
+    renderBorder(this.ctx, this.dimensions, STYLE.tracks.edgeColor);
     drawDotsScaled(this.ctx, dots, xScale, yScale);
   }
 }