+++ assets/js/gens.ts
@@ -18,8 +18,8 @@ export {
   queryRegionOrGene,
 } from "./navigation";
 
-import "./components/genstracks";
-import { GensTracks } from "./components/genstracks";
+import "./components/tracks_manager";
+import { TracksManager } from "./components/tracks_manager";
 import "./components/input_controls";
 import { InputControls } from "./components/input_controls";
 import {
@@ -61,7 +61,7 @@ export async function initCanvases({
   annotationFile: string;
   startRegion: Region;
 }) {
-  const gensTracks = document.getElementById("gens-tracks") as GensTracks;
+  const gensTracks = document.getElementById("gens-tracks") as TracksManager;
 
   const inputControls = document.getElementById(
     "input-controls",
@@ -69,100 +69,136 @@ export async function initCanvases({
 
   const api = new GensAPI(sampleId, caseId, genomeBuild, gensApiURL);
 
+  const renderDataSource = getRenderDataSource(
+    api,
+    () => {
+      const region = inputControls.getRegion();
+      return region.chrom;
+    },
+  );
+
   const allChromData = await api.getAllChromData();
 
   const onChromClick = async (chrom) => {
     const chromData = await api.getChromData(chrom);
     inputControls.updateChromosome(chrom, chromData.size);
-    const selectedAnnots = inputControls.getAnnotations();
-    const region = inputControls.getRegion();
-
-    const renderData = await parseRenderData(
-      api,
-      startRegion.chrom,
-      selectedAnnots,
-    );
-    gensTracks.updateRenderData(renderData, region);
-    gensTracks.render();
+    // await gensTracks.updateRenderData();
+    const updateData = true;
+    gensTracks.render(updateData);
   };
 
-  gensTracks.initialize(allChromData, onChromClick);
+  const getChromInfo = (chromosome: string) => {
+    return allChromData[chromosome];
+  };
 
-  const renderData = await parseRenderData(api, startRegion.chrom, [
-    annotationFile,
-  ]);
-  gensTracks.updateRenderData(renderData, startRegion);
-  gensTracks.render();
-
-  // FIXME: Look into how to parse this for predefined start URLs
-  inputControls.initialize(
+  initialize(
+    inputControls,
+    gensTracks,
     startRegion,
-    [annotationFile],
-    async (region) => {},
-    async (region, _source) => {
-      const selectedAnnots = inputControls.getAnnotations();
-      const renderData = await parseRenderData(
-        api,
-        region.chrom,
-        selectedAnnots,
-      );
-      gensTracks.updateRenderData(renderData, region);
-      gensTracks.render();
-    },
-    async (_newXRange) => {
-      const selectedAnnots = inputControls.getAnnotations();
-      const region = inputControls.getRegion();
-      const renderData = await parseRenderData(
-        api,
-        region.chrom,
-        selectedAnnots,
-      );
-      gensTracks.updateRenderData(renderData, region);
-      gensTracks.render();
-    },
+    annotationFile,
     gensApiURL,
+    onChromClick,
+    getChromInfo,
+    renderDataSource,
   );
 }
 
-async function parseRenderData(
-  gensDb: GensAPI,
-  chrom: string,
-  annotSources: string[],
-): Promise<RenderData> {
-
-  const parsedAnnotationData = {};
-  for (const source of annotSources) {
-    const annotData = await gensDb.getAnnotations(chrom, source);
-    parsedAnnotationData[source] = parseAnnotations(annotData);
-  }
-
-  const overviewCovRaw = await gensDb.getOverviewCovData();
-  const overviewCovRender = transformMap(overviewCovRaw, (cov) =>
-    parseCoverageDot(cov, STYLE.colors.darkGray),
+async function initialize(
+  inputControls: InputControls,
+  tracks: TracksManager,
+  startRegion: Region,
+  defaultAnnotation: string,
+  gensApiURI: string,
+  onChromClick: (string) => void,
+  getChromInfo: (string) => ChromosomeInfo,
+  renderDataSource: RenderDataSource,
+) {
+  // FIXME: Look into how to parse this for predefined start URLs
+  inputControls.initialize(
+    startRegion,
+    [defaultAnnotation],
+    async (_region, _source) => {
+      tracks.render(true);
+    },
+    async (_newXRange) => {
+      tracks.render(true);
+    },
+    gensApiURI,
   );
-  const overviewBafRaw = await gensDb.getOverviewBafData();
-  const overviewBafRender = transformMap(overviewBafRaw, (cov) =>
-    parseCoverageDot(cov, STYLE.colors.darkGray),
+
+  await tracks.initialize(
+    getChromInfo,
+    onChromClick,
+    renderDataSource,
+    () => inputControls.getRegion().chrom,
+    () => inputControls.getRange(),
+    () => inputControls.getAnnotSources(),
   );
 
-  const covRaw = await gensDb.getCov(chrom);
-  // const covData = parseCoverageBin(covRaw)
+  // await tracks.updateRenderData();
+  tracks.render(true);
+}
 
-  const bafRaw = await gensDb.getBaf(chrom);
-  const transcriptsRaw = await gensDb.getTranscripts(chrom);
-  const variantsRaw = await gensDb.getVariants(chrom);
-
-  const renderData: RenderData = {
-    chromInfo: await gensDb.getChromData(chrom),
-    annotations: parsedAnnotationData,
-    covData: parseCoverageDot(covRaw, STYLE.colors.teal),
-    bafData: parseCoverageDot(bafRaw, STYLE.colors.orange),
-    transcriptData: parseTranscripts(transcriptsRaw),
-    variantData: parseVariants(variantsRaw, STYLE.colors.variantColors),
-    overviewCovData: overviewCovRender,
-    overviewBafData: overviewBafRender,
+function getRenderDataSource(
+  gensAPI: GensAPI,
+  getChrom: () => string,
+): RenderDataSource {
+  const getChromInfo = async () => {
+    return await gensAPI.getChromData(getChrom());
   };
-  return renderData;
+
+  const getAnnotation = async (source: string) => {
+    const annotData = await gensAPI.getAnnotations(getChrom(), source);
+    return parseAnnotations(annotData);
+  };
+
+  const getCovData = async () => {
+    const covRaw = await gensAPI.getCov(getChrom());
+    return parseCoverageDot(covRaw, STYLE.colors.teal);
+  };
+
+  const getBafData = async () => {
+    const bafRaw = await gensAPI.getBaf(getChrom());
+    return parseCoverageDot(bafRaw, STYLE.colors.orange);
+  };
+
+  const getTranscriptData = async () => {
+    const transcriptsRaw = await gensAPI.getTranscripts(getChrom());
+    return parseTranscripts(transcriptsRaw);
+  };
+
+  const getVariantData = async () => {
+    const variantsRaw = await gensAPI.getVariants(getChrom());
+    return parseVariants(variantsRaw, STYLE.colors.variantColors);
+  };
+
+  const getOverviewCovData = async () => {
+    const overviewCovRaw = await gensAPI.getOverviewCovData();
+    const overviewCovRender = transformMap(overviewCovRaw, (cov) =>
+      parseCoverageDot(cov, STYLE.colors.darkGray),
+    );
+    return overviewCovRender;
+  };
+
+  const getOverviewBafData = async () => {
+    const overviewBafRaw = await gensAPI.getOverviewBafData();
+    const overviewBafRender = transformMap(overviewBafRaw, (cov) =>
+      parseCoverageDot(cov, STYLE.colors.darkGray),
+    );
+    return overviewBafRender;
+  };
+
+  const renderDataSource: RenderDataSource = {
+    getChromInfo,
+    getAnnotation,
+    getCovData,
+    getBafData,
+    getTranscriptData,
+    getVariantData,
+    getOverviewCovData,
+    getOverviewBafData,
+  };
+  return renderDataSource;
 }
 
 // Make hard link and copy link to clipboard